Move animated checkmark to its own component
