+++ src/components/GettingStartedTour/TourPopup.scss
@@ -351,58 +351,3 @@ $arrows-line-width: 2px;
     }
 }
 
-/* Animated checkmark */
-$color : #6fbf39;
-$stroke-width: 5px;
-
-.checkmark__circle {
-    stroke-dasharray: 166;
-    stroke-dashoffset: 166;
-    stroke-width: $stroke-width;
-    stroke-miterlimit: 10;
-    stroke: $color;
-    fill: none;
-    animation: stroke 0.6s cubic-bezier(0.65, 0, 0.45, 1) forwards;
-}
-
-.checkmark {
-    width: 100%;
-    height: 100%;
-    border-radius: 50%;
-    stroke-width: 5;
-    stroke: white;
-    stroke-miterlimit: 10;
-    box-shadow: inset 0px 0px 0px $color;
-    animation: fill .4s ease-in-out .4s forwards, scale .3s ease-in-out .9s both;
-}
-
-.checkmark__check {
-    transform-origin: 50% 50%;
-    stroke-dasharray: 48;
-    stroke-dashoffset: 48;
-    animation: stroke 0.3s cubic-bezier(0.65, 0, 0.45, 1) 0.8s forwards;
-}
-
-@keyframes stroke {
-    100% {
-        stroke-dashoffset: 0;
-    }
-}
-
-@keyframes scale {
-    0%,
-    100% {
-        transform: none;
-    }
-
-    50% {
-        transform: scale3d(1.1, 1.1, 1);
-    }
-}
-
-@keyframes fill {
-    100% {
-        box-shadow: inset 0px 0px 0px 1000px $color;
-    }
-}
-
